Move the API accesses to a separate file
